ComputeSimulator - defining an executable
